Chore/ANC tokens storage improvement (#158)
* Standalone OAuthContainer file.

* Update logic.
